reference/original-sources/DeepResearchAgent-main/pyproject.toml
[project]
name = "deepresearchagent"
version = "0.1.0"
description = "DeepResearchAgent - A hierarchical multi-agent framework for general-purpose task solving"
authors = [
    {name = "Skywork", email = "info@skywork.ai"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv",
    "rich",
    "litellm",
    "openai",
    "pillow",
    "numpy>=1.24.0", # Relaxed to permit numpy 2.x (align with requirements.txt). If you need to block numpy 2.x for compatibility, revert this change.
    "pynput",
    "transformers",
    "markdownify",
    "duckduckgo-search",
    "mammoth",
    "pandas",
    "pdfminer-six",
    "python-pptx",
    "puremagic",
    "pydub",
    "speechrecognition",
    "pathvalidate",
    "serpapi",
    "google-search-results",
    "datasets",
    "anthropic",
    "openpyxl",
    "baidusearch",
    "patchright",
    "pyautogui",
    "json5",
    "tenacity",
    "crawl4ai",
    "camelot-py",
    "xlrd",
    "toml",
    "googlesearch-python",
    "fastmcp",
    "httpx-aiohttp",
    "inflection",
    "mmengine",
    "pydantic>=2.10.4,<3.0.0", # Updated for browser-use compatibility
    "markitdown", # Document converter
    "browser-use==0.1.48", # Keep specific version for compatibility
    "playwright",
    "firecrawl-py>=4.3.6",
    "langchain",
    "langchain-openai",
    "langchain-anthropic",
    "langchain-google-genai",
    "pytesseract>=0.3.13",
    "coolprop>=7.0.0",
    "textual>=6.1.0",
    "cairosvg>=2.8.2",
    "pymupdf>=1.26.4",
    "pdfplumber>=0.11.7",
    "pillow-heif>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
    "mypy",
    "ruff",  # Modern, fast linter alternative to flake8
]

[project.scripts]
deepresearch = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "black",
    "isort",
    "ruff",
    "mypy",
]

[tool.black]
line-length = 88
target-version = ['py311', 'py312', 'py313']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by black)

[tool.ruff.lint.isort]
known-first-party = ["src"]
